+++ primitives/src/merkle_tree/internal.rs
@@ -8,10 +8,9 @@ use super::{
     DigestAlgorithm, Element, Index, LookupResult, MerkleCommitment, NodeValue, ToTraversalPath,
 };
 use crate::errors::{PrimitivesError, VerificationResult};
+use alloc::sync::Arc;
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use ark_std::{
-    borrow::Borrow, boxed::Box, format, iter::Peekable, string::ToString, vec, vec::Vec,
-};
+use ark_std::{borrow::Borrow, format, iter::Peekable, string::ToString, vec, vec::Vec};
 use core::marker::PhantomData;
 use itertools::Itertools;
 use jf_utils::canonical;
@@ -28,7 +27,7 @@ pub enum MerkleNode<E: Element, I: Index, T: NodeValue> {
     Branch {
         #[serde(with = "canonical")]
         value: T,
-        children: Vec<Box<MerkleNode<E, I, T>>>,
+        children: Vec<Arc<MerkleNode<E, I, T>>>,
     },
     Leaf {
         #[serde(with = "canonical")]
@@ -177,7 +176,7 @@ where
 pub(crate) fn build_tree_internal<E, H, Arity, T>(
     height: Option<usize>,
     elems: impl IntoIterator<Item = impl Borrow<E>>,
-) -> Result<(Box<MerkleNode<E, u64, T>>, usize, u64), PrimitivesError>
+) -> Result<(Arc<MerkleNode<E, u64, T>>, usize, u64), PrimitivesError>
 where
     E: Element,
     H: DigestAlgorithm<E, u64, T>,
@@ -211,15 +210,15 @@ where
                 let children = chunk
                     .map(|(pos, elem)| {
                         let pos = pos as u64;
-                        Ok(Box::new(MerkleNode::Leaf {
+                        Ok(Arc::new(MerkleNode::Leaf {
                             value: H::digest_leaf(&pos, elem.borrow())?,
                             pos,
                             elem: elem.borrow().clone(),
                         }))
                     })
-                    .pad_using(Arity::to_usize(), |_| Ok(Box::new(MerkleNode::Empty)))
+                    .pad_using(Arity::to_usize(), |_| Ok(Arc::new(MerkleNode::Empty)))
                     .collect::<Result<Vec<_>, PrimitivesError>>()?;
-                Ok(Box::new(MerkleNode::<E, u64, T>::Branch {
+                Ok(Arc::new(MerkleNode::<E, u64, T>::Branch {
                     value: digest_branch::<E, H, u64, T>(&children)?,
                     children,
                 }))
@@ -233,10 +232,10 @@ where
                 .map(|chunk| {
                     let children = chunk
                         .pad_using(Arity::to_usize(), |_| {
-                            Box::new(MerkleNode::<E, u64, T>::Empty)
+                            Arc::new(MerkleNode::<E, u64, T>::Empty)
                         })
                         .collect::<Vec<_>>();
-                    Ok(Box::new(MerkleNode::<E, u64, T>::Branch {
+                    Ok(Arc::new(MerkleNode::<E, u64, T>::Branch {
                         value: digest_branch::<E, H, u64, T>(&children)?,
                         children,
                     }))
@@ -245,7 +244,7 @@ where
         }
         Ok((cur_nodes[0].clone(), height, num_leaves))
     } else {
-        Ok((Box::new(MerkleNode::<E, u64, T>::Empty), height, 0))
+        Ok((Arc::new(MerkleNode::<E, u64, T>::Empty), height, 0))
     }
 }
 
@@ -253,7 +252,7 @@ where
 pub(crate) fn build_light_weight_tree_internal<E, H, Arity, T>(
     height: Option<usize>,
     elems: impl IntoIterator<Item = impl Borrow<E>>,
-) -> Result<(Box<MerkleNode<E, u64, T>>, usize, u64), PrimitivesError>
+) -> Result<(Arc<MerkleNode<E, u64, T>>, usize, u64), PrimitivesError>
 where
     E: Element,
     H: DigestAlgorithm<E, u64, T>,
@@ -290,27 +289,27 @@ where
                     .map(|(pos, elem)| {
                         let pos = pos as u64;
                         Ok(if pos < num_leaves - 1 {
-                            Box::new(MerkleNode::ForgettenSubtree {
+                            Arc::new(MerkleNode::ForgettenSubtree {
                                 value: H::digest_leaf(&pos, elem.borrow())?,
                             })
                         } else {
-                            Box::new(MerkleNode::Leaf {
+                            Arc::new(MerkleNode::Leaf {
                                 value: H::digest_leaf(&pos, elem.borrow())?,
                                 pos,
                                 elem: elem.borrow().clone(),
                             })
                         })
                     })
-                    .pad_using(Arity::to_usize(), |_| Ok(Box::new(MerkleNode::Empty)))
+                    .pad_using(Arity::to_usize(), |_| Ok(Arc::new(MerkleNode::Empty)))
                     .collect::<Result<Vec<_>, PrimitivesError>>()?;
-                Ok(Box::new(MerkleNode::<E, u64, T>::Branch {
+                Ok(Arc::new(MerkleNode::<E, u64, T>::Branch {
                     value: digest_branch::<E, H, u64, T>(&children)?,
                     children,
                 }))
             })
             .collect::<Result<Vec<_>, PrimitivesError>>()?;
         for i in 1..cur_nodes.len() - 1 {
-            cur_nodes[i] = Box::new(MerkleNode::ForgettenSubtree {
+            cur_nodes[i] = Arc::new(MerkleNode::ForgettenSubtree {
                 value: cur_nodes[i].value(),
             })
         }
@@ -322,29 +321,29 @@ where
                 .map(|chunk| {
                     let children = chunk
                         .pad_using(Arity::to_usize(), |_| {
-                            Box::new(MerkleNode::<E, u64, T>::Empty)
+                            Arc::new(MerkleNode::<E, u64, T>::Empty)
                         })
                         .collect::<Vec<_>>();
-                    Ok(Box::new(MerkleNode::<E, u64, T>::Branch {
+                    Ok(Arc::new(MerkleNode::<E, u64, T>::Branch {
                         value: digest_branch::<E, H, u64, T>(&children)?,
                         children,
                     }))
                 })
                 .collect::<Result<Vec<_>, PrimitivesError>>()?;
             for i in 1..cur_nodes.len() - 1 {
-                cur_nodes[i] = Box::new(MerkleNode::ForgettenSubtree {
+                cur_nodes[i] = Arc::new(MerkleNode::ForgettenSubtree {
                     value: cur_nodes[i].value(),
                 })
             }
         }
         Ok((cur_nodes[0].clone(), height, num_leaves))
     } else {
-        Ok((Box::new(MerkleNode::<E, u64, T>::Empty), height, 0))
+        Ok((Arc::new(MerkleNode::<E, u64, T>::Empty), height, 0))
     }
 }
 
 pub(crate) fn digest_branch<E, H, I, T>(
-    data: &[Box<MerkleNode<E, I, T>>],
+    data: &[Arc<MerkleNode<E, I, T>>],
 ) -> Result<T, PrimitivesError>
 where
     E: Element,
@@ -365,17 +364,26 @@ where
 {
     /// Forget a leaf from the merkle tree. Internal branch merkle node will
     /// also be forgotten if all its leafs are forgotten.
+    /// WARN(#495): this method breaks non-membership proofs.
+    #[allow(clippy::type_complexity)]
     pub(crate) fn forget_internal(
-        &mut self,
+        &self,
         height: usize,
         traversal_path: &[usize],
-    ) -> LookupResult<E, MerklePath<E, I, T>, MerklePath<E, I, T>> {
+    ) -> (
+        Arc<Self>,
+        LookupResult<E, MerklePath<E, I, T>, MerklePath<E, I, T>>,
+    ) {
         match self {
-            MerkleNode::Empty => LookupResult::NotFound(vec![MerkleNode::Empty; height + 1]),
+            MerkleNode::Empty => (
+                Arc::new(self.clone()),
+                LookupResult::NotFound(vec![MerkleNode::Empty; height + 1]),
+            ),
             MerkleNode::Branch { value, children } => {
-                match children[traversal_path[height - 1]]
-                    .forget_internal(height - 1, traversal_path)
-                {
+                let mut children = children.clone();
+                let (new_child, result) = children[traversal_path[height - 1]]
+                    .forget_internal(height - 1, traversal_path);
+                match result {
                     LookupResult::Ok(elem, mut proof) => {
                         proof.push(MerkleNode::Branch {
                             value: T::default(),
@@ -383,26 +391,39 @@ where
                                 .iter()
                                 .map(|child| {
                                     if let MerkleNode::Empty = **child {
-                                        Box::new(MerkleNode::Empty)
+                                        Arc::new(MerkleNode::Empty)
                                     } else {
-                                        Box::new(MerkleNode::ForgettenSubtree {
+                                        Arc::new(MerkleNode::ForgettenSubtree {
                                             value: child.value(),
                                         })
                                     }
                                 })
                                 .collect::<Vec<_>>(),
                         });
+                        children[traversal_path[height - 1]] = new_child;
                         if children.iter().all(|child| {
                             matches!(
                                 **child,
-                                MerkleNode::Empty | MerkleNode::ForgettenSubtree { value: _ }
+                                MerkleNode::Empty | MerkleNode::ForgettenSubtree { .. }
                             )
                         }) {
-                            *self = MerkleNode::ForgettenSubtree { value: *value };
+                            (
+                                Arc::new(MerkleNode::ForgettenSubtree { value: *value }),
+                                LookupResult::Ok(elem, proof),
+                            )
+                        } else {
+                            (
+                                Arc::new(MerkleNode::Branch {
+                                    value: *value,
+                                    children,
+                                }),
+                                LookupResult::Ok(elem, proof),
+                            )
                         }
-                        LookupResult::Ok(elem, proof)
                     },
-                    LookupResult::NotInMemory => LookupResult::NotInMemory,
+                    LookupResult::NotInMemory => {
+                        (Arc::new(self.clone()), LookupResult::NotInMemory)
+                    },
                     LookupResult::NotFound(mut non_membership_proof) => {
                         non_membership_proof.push(MerkleNode::Branch {
                             value: T::default(),
@@ -410,16 +431,19 @@ where
                                 .iter()
                                 .map(|child| {
                                     if let MerkleNode::Empty = **child {
-                                        Box::new(MerkleNode::Empty)
+                                        Arc::new(MerkleNode::Empty)
                                     } else {
-                                        Box::new(MerkleNode::ForgettenSubtree {
+                                        Arc::new(MerkleNode::ForgettenSubtree {
                                             value: child.value(),
                                         })
                                     }
                                 })
                                 .collect::<Vec<_>>(),
                         });
-                        LookupResult::NotFound(non_membership_proof)
+                        (
+                            Arc::new(self.clone()),
+                            LookupResult::NotFound(non_membership_proof),
+                        )
                     },
                 }
             },
@@ -430,21 +454,24 @@ where
                     pos: pos.clone(),
                     elem: elem.clone(),
                 }];
-                *self = MerkleNode::ForgettenSubtree { value: *value };
-                LookupResult::Ok(elem, proof)
+                (
+                    Arc::new(MerkleNode::ForgettenSubtree { value: *value }),
+                    LookupResult::Ok(elem, proof),
+                )
             },
-            _ => LookupResult::NotInMemory,
+            _ => (Arc::new(self.clone()), LookupResult::NotInMemory),
         }
     }
 
-    /// Re-insert a forgotten leaf to the Merkle tree if the proof is valid.
+    /// Re-insert a forgotten leaf to the Merkle tree. We assume that the proof
+    /// is valid and already checked.
     pub(crate) fn remember_internal<H, Arity>(
-        &mut self,
+        &self,
         height: usize,
         traversal_path: &[usize],
         path_values: &[T],
         proof: &[MerkleNode<E, I, T>],
-    ) -> Result<(), PrimitivesError>
+    ) -> Result<Arc<Self>, PrimitivesError>
     where
         H: DigestAlgorithm<E, I, T>,
         Arity: Unsigned,
@@ -458,36 +485,46 @@ where
             )));
         }
 
-        match (&mut *self, &proof[height]) {
+        match (self, &proof[height]) {
             (Self::ForgettenSubtree { value }, Self::Branch { children, .. }) => {
                 // Recurse into the appropriate sub-tree to remember the rest of the path.
                 let mut children = children.clone();
-                children[traversal_path[height - 1]].remember_internal::<H, Arity>(
-                    height - 1,
-                    traversal_path,
-                    path_values,
-                    proof,
-                )?;
+                children[traversal_path[height - 1]] = children[traversal_path[height - 1]]
+                    .remember_internal::<H, Arity>(
+                        height - 1,
+                        traversal_path,
+                        path_values,
+                        proof,
+                    )?;
                 // Remember `*self`.
-                *self = Self::Branch {
+                Ok(Arc::new(Self::Branch {
                     value: *value,
                     children,
-                };
-                Ok(())
+                }))
             },
             (Self::ForgettenSubtree { .. }, node) => {
                 // Replace forgotten sub-tree with a hopefully-less-forgotten sub-tree from the
                 // proof. Safe because we already checked our hash value matches the proof.
-                *self = node.clone();
-                Ok(())
+                Ok(Arc::new(node.clone()))
             },
-            (Self::Branch { children, .. }, Self::Branch { .. }) => children
-                [traversal_path[height - 1]]
-                .remember_internal::<H, Arity>(height - 1, traversal_path, path_values, proof),
+            (Self::Branch { value, children }, Self::Branch { .. }) => {
+                let mut children = children.clone();
+                children[traversal_path[height - 1]] = children[traversal_path[height - 1]]
+                    .remember_internal::<H, Arity>(
+                        height - 1,
+                        traversal_path,
+                        path_values,
+                        proof,
+                    )?;
+                Ok(Arc::new(Self::Branch {
+                    value: *value,
+                    children,
+                }))
+            },
             (Self::Leaf { .. }, Self::Leaf { .. }) | (Self::Empty, Self::Empty) => {
                 // This node is already a complete sub-tree, so there's nothing to remember. The
                 // proof matches, so just return success.
-                Ok(())
+                Ok(Arc::new(self.clone()))
             },
             (..) => Err(PrimitivesError::ParameterError("Invalid proof".into())),
         }
@@ -517,9 +554,9 @@ where
                                 .iter()
                                 .map(|child| {
                                     if let MerkleNode::Empty = **child {
-                                        Box::new(MerkleNode::Empty)
+                                        Arc::new(MerkleNode::Empty)
                                     } else {
-                                        Box::new(MerkleNode::ForgettenSubtree {
+                                        Arc::new(MerkleNode::ForgettenSubtree {
                                             value: child.value(),
                                         })
                                     }
@@ -536,9 +573,9 @@ where
                                 .iter()
                                 .map(|child| {
                                     if let MerkleNode::Empty = **child {
-                                        Box::new(MerkleNode::Empty)
+                                        Arc::new(MerkleNode::Empty)
                                     } else {
-                                        Box::new(MerkleNode::ForgettenSubtree {
+                                        Arc::new(MerkleNode::ForgettenSubtree {
                                             value: child.value(),
                                         })
                                     }
@@ -563,13 +600,14 @@ where
     ///   result)`, `delta` represents the changes to the overall number of
     ///   leaves of the tree, `result` contains the original lookup information
     ///   at the given location.
+    #[allow(clippy::type_complexity)]
     pub(crate) fn update_with_internal<H, Arity, F>(
-        &mut self,
+        &self,
         height: usize,
         pos: impl Borrow<I>,
         traversal_path: &[usize],
         f: F,
-    ) -> Result<(i64, LookupResult<E, (), ()>), PrimitivesError>
+    ) -> Result<(Arc<Self>, i64, LookupResult<E, (), ()>), PrimitivesError>
     where
         H: DigestAlgorithm<E, I, T>,
         Arity: Unsigned,
@@ -579,20 +617,21 @@ where
         match self {
             MerkleNode::Leaf {
                 elem: node_elem,
-                value,
+                value: _,
                 pos,
             } => {
                 let result = LookupResult::Ok(node_elem.clone(), ());
                 match f(Some(node_elem)) {
-                    Some(elem) => {
-                        *value = H::digest_leaf(pos, &elem)?;
-                        *node_elem = elem;
-                        Ok((0i64, result))
-                    },
-                    None => {
-                        *self = MerkleNode::Empty;
-                        Ok((-1i64, result))
-                    },
+                    Some(elem) => Ok((
+                        Arc::new(MerkleNode::Leaf {
+                            value: H::digest_leaf(pos, &elem)?,
+                            pos: pos.clone(),
+                            elem,
+                        }),
+                        0i64,
+                        result,
+                    )),
+                    None => Ok((Arc::new(MerkleNode::Empty), -1i64, result)),
                 }
             },
             MerkleNode::Branch { value, children } => {
@@ -603,56 +642,89 @@ where
                     traversal_path,
                     f,
                 )?;
+                let mut children = children.clone();
+                children[branch] = result.0;
                 if matches!(*children[branch], MerkleNode::ForgettenSubtree { .. }) {
                     // If the branch containing the update was forgotten by
                     // user, the update failed and nothing was changed, so we
                     // can short-circuit without recomputing this node's value.
+                    Ok((
+                        Arc::new(MerkleNode::Branch {
+                            value: *value,
+                            children,
+                        }),
+                        result.1,
+                        result.2,
+                    ))
                 } else if children
                     .iter()
                     .all(|child| matches!(**child, MerkleNode::Empty))
                 {
-                    // If all children are empty, remove the current node.
-                    *self = MerkleNode::Empty;
+                    Ok((Arc::new(MerkleNode::Empty), result.1, result.2))
                 } else {
                     // Otherwise, an entry has been updated and the value of one of our children has
                     // changed, so we must recompute our own value.
-                    *value = digest_branch::<E, H, I, T>(children)?;
+                    // *value = digest_branch::<E, H, I, T>(&children)?;
+                    Ok((
+                        Arc::new(MerkleNode::Branch {
+                            value: digest_branch::<E, H, I, T>(&children)?,
+                            children,
+                        }),
+                        result.1,
+                        result.2,
+                    ))
                 }
-                Ok(result)
             },
             MerkleNode::Empty => {
                 if height == 0 {
                     if let Some(elem) = f(None) {
-                        *self = MerkleNode::Leaf {
-                            value: H::digest_leaf(pos, &elem)?,
-                            pos: pos.clone(),
-                            elem,
-                        };
-                        Ok((1, LookupResult::NotFound(())))
+                        Ok((
+                            Arc::new(MerkleNode::Leaf {
+                                value: H::digest_leaf(pos, &elem)?,
+                                pos: pos.clone(),
+                                elem,
+                            }),
+                            1i64,
+                            LookupResult::NotFound(()),
+                        ))
                     } else {
-                        Ok((0, LookupResult::NotFound(())))
+                        Ok((
+                            Arc::new(MerkleNode::Empty),
+                            0i64,
+                            LookupResult::NotFound(()),
+                        ))
                     }
                 } else {
                     let branch = traversal_path[height - 1];
-                    let mut children = vec![Box::new(MerkleNode::Empty); Arity::to_usize()];
+                    let mut children = (0..Arity::to_usize())
+                        .map(|_| Arc::new(Self::Empty))
+                        .collect::<Vec<_>>();
+                    // Inserting new leave here, shortcutting
                     let result = children[branch].update_with_internal::<H, Arity, _>(
                         height - 1,
                         pos,
                         traversal_path,
                         f,
                     )?;
+                    children[branch] = result.0;
                     if matches!(*children[branch], MerkleNode::Empty) {
                         // No update performed.
+                        Ok((Arc::new(MerkleNode::Empty), 0i64, result.2))
                     } else {
-                        *self = MerkleNode::Branch {
-                            value: digest_branch::<E, H, I, T>(&children)?,
-                            children,
-                        };
+                        Ok((
+                            Arc::new(MerkleNode::Branch {
+                                value: digest_branch::<E, H, I, T>(&children)?,
+                                children,
+                            }),
+                            result.1,
+                            result.2,
+                        ))
                     }
-                    Ok(result)
                 }
             },
-            MerkleNode::ForgettenSubtree { .. } => Ok((0, LookupResult::NotInMemory)),
+            MerkleNode::ForgettenSubtree { .. } => Err(PrimitivesError::InternalError(
+                "Given leaf is not in memory".to_string(),
+            )),
         }
     }
 }
@@ -664,23 +736,23 @@ where
 {
     /// Batch insertion for the given Merkle node.
     pub(crate) fn extend_internal<H, Arity>(
-        &mut self,
+        &self,
         height: usize,
         pos: &u64,
         traversal_path: &[usize],
         at_frontier: bool,
         data: &mut Peekable<impl Iterator<Item = impl Borrow<E>>>,
-    ) -> Result<u64, PrimitivesError>
+    ) -> Result<(Arc<Self>, u64), PrimitivesError>
     where
         H: DigestAlgorithm<E, u64, T>,
         Arity: Unsigned,
     {
         if data.peek().is_none() {
-            return Ok(0);
+            return Ok((Arc::new(self.clone()), 0));
         }
         let mut cur_pos = *pos;
         match self {
-            MerkleNode::Branch { value, children } => {
+            MerkleNode::Branch { value: _, children } => {
                 let mut cnt = 0u64;
                 let mut frontier = if at_frontier {
                     traversal_path[height - 1]
@@ -688,31 +760,35 @@ where
                     0
                 };
                 let cap = Arity::to_usize();
+                let mut children = children.clone();
                 while data.peek().is_some() && frontier < cap {
-                    let increment = children[frontier].extend_internal::<H, Arity>(
+                    let (new_child, increment) = children[frontier].extend_internal::<H, Arity>(
                         height - 1,
                         &cur_pos,
                         traversal_path,
                         at_frontier && frontier == traversal_path[height - 1],
                         data,
                     )?;
+                    children[frontier] = new_child;
                     cnt += increment;
                     cur_pos += increment;
                     frontier += 1;
                 }
-                *value = digest_branch::<E, H, u64, T>(children)?;
-                Ok(cnt)
+                let value = digest_branch::<E, H, u64, T>(&children)?;
+                Ok((Arc::new(Self::Branch { value, children }), cnt))
             },
             MerkleNode::Empty => {
                 if height == 0 {
                     let elem = data.next().unwrap();
                     let elem = elem.borrow();
-                    *self = MerkleNode::Leaf {
-                        value: H::digest_leaf(pos, elem)?,
-                        pos: *pos,
-                        elem: elem.clone(),
-                    };
-                    Ok(1)
+                    Ok((
+                        Arc::new(MerkleNode::Leaf {
+                            value: H::digest_leaf(pos, elem)?,
+                            pos: *pos,
+                            elem: elem.clone(),
+                        }),
+                        1,
+                    ))
                 } else {
                     let mut cnt = 0u64;
                     let mut frontier = if at_frontier {
@@ -721,24 +797,28 @@ where
                         0
                     };
                     let cap = Arity::to_usize();
-                    let mut children = vec![Box::new(MerkleNode::Empty); cap];
+                    let mut children = (0..cap).map(|_| Arc::new(Self::Empty)).collect::<Vec<_>>();
                     while data.peek().is_some() && frontier < cap {
-                        let increment = children[frontier].extend_internal::<H, Arity>(
-                            height - 1,
-                            &cur_pos,
-                            traversal_path,
-                            at_frontier && frontier == traversal_path[height - 1],
-                            data,
-                        )?;
+                        let (new_child, increment) = children[frontier]
+                            .extend_internal::<H, Arity>(
+                                height - 1,
+                                &cur_pos,
+                                traversal_path,
+                                at_frontier && frontier == traversal_path[height - 1],
+                                data,
+                            )?;
+                        children[frontier] = new_child;
                         cnt += increment;
                         cur_pos += increment;
                         frontier += 1;
                     }
-                    *self = MerkleNode::Branch {
-                        value: digest_branch::<E, H, u64, T>(&children)?,
-                        children,
-                    };
-                    Ok(cnt)
+                    Ok((
+                        Arc::new(MerkleNode::Branch {
+                            value: digest_branch::<E, H, u64, T>(&children)?,
+                            children,
+                        }),
+                        cnt,
+                    ))
                 }
             },
             MerkleNode::Leaf { .. } => Err(PrimitivesError::ParameterError(
@@ -753,23 +833,23 @@ where
     /// Similar to [`extend_internal`], but this function will automatically
     /// forget every leaf except for the Merkle tree frontier.
     pub(crate) fn extend_and_forget_internal<H, Arity>(
-        &mut self,
+        &self,
         height: usize,
         pos: &u64,
         traversal_path: &[usize],
         at_frontier: bool,
         data: &mut Peekable<impl Iterator<Item = impl Borrow<E>>>,
-    ) -> Result<u64, PrimitivesError>
+    ) -> Result<(Arc<Self>, u64), PrimitivesError>
     where
         H: DigestAlgorithm<E, u64, T>,
         Arity: Unsigned,
     {
         if data.peek().is_none() {
-            return Ok(0);
+            return Ok((Arc::new(self.clone()), 0));
         }
         let mut cur_pos = *pos;
         match self {
-            MerkleNode::Branch { value, children } => {
+            MerkleNode::Branch { value: _, children } => {
                 let mut cnt = 0u64;
                 let mut frontier = if at_frontier {
                     traversal_path[height - 1]
@@ -777,37 +857,42 @@ where
                     0
                 };
                 let cap = Arity::to_usize();
+                let mut children = children.clone();
                 while data.peek().is_some() && frontier < cap {
                     if frontier > 0 && !children[frontier - 1].is_forgotten() {
                         children[frontier - 1] =
-                            Box::new(MerkleNode::<E, u64, T>::ForgettenSubtree {
+                            Arc::new(MerkleNode::<E, u64, T>::ForgettenSubtree {
                                 value: children[frontier - 1].value(),
                             });
                     }
-                    let increment = children[frontier].extend_and_forget_internal::<H, Arity>(
-                        height - 1,
-                        &cur_pos,
-                        traversal_path,
-                        at_frontier && frontier == traversal_path[height - 1],
-                        data,
-                    )?;
+                    let (new_child, increment) = children[frontier]
+                        .extend_and_forget_internal::<H, Arity>(
+                            height - 1,
+                            &cur_pos,
+                            traversal_path,
+                            at_frontier && frontier == traversal_path[height - 1],
+                            data,
+                        )?;
+                    children[frontier] = new_child;
                     cnt += increment;
                     cur_pos += increment;
                     frontier += 1;
                 }
-                *value = digest_branch::<E, H, u64, T>(children)?;
-                Ok(cnt)
+                let value = digest_branch::<E, H, u64, T>(&children)?;
+                Ok((Arc::new(Self::Branch { value, children }), cnt))
             },
             MerkleNode::Empty => {
                 if height == 0 {
                     let elem = data.next().unwrap();
                     let elem = elem.borrow();
-                    *self = MerkleNode::Leaf {
-                        value: H::digest_leaf(pos, elem)?,
-                        pos: *pos,
-                        elem: elem.clone(),
-                    };
-                    Ok(1)
+                    Ok((
+                        Arc::new(MerkleNode::Leaf {
+                            value: H::digest_leaf(pos, elem)?,
+                            pos: *pos,
+                            elem: elem.clone(),
+                        }),
+                        1,
+                    ))
                 } else {
                     let mut cnt = 0u64;
                     let mut frontier = if at_frontier {
@@ -816,30 +901,34 @@ where
                         0
                     };
                     let cap = Arity::to_usize();
-                    let mut children = vec![Box::new(MerkleNode::Empty); cap];
+                    let mut children = (0..cap).map(|_| Arc::new(Self::Empty)).collect::<Vec<_>>();
                     while data.peek().is_some() && frontier < cap {
                         if frontier > 0 && !children[frontier - 1].is_forgotten() {
                             children[frontier - 1] =
-                                Box::new(MerkleNode::<E, u64, T>::ForgettenSubtree {
+                                Arc::new(MerkleNode::<E, u64, T>::ForgettenSubtree {
                                     value: children[frontier - 1].value(),
                                 });
                         }
-                        let increment = children[frontier].extend_and_forget_internal::<H, Arity>(
-                            height - 1,
-                            &cur_pos,
-                            traversal_path,
-                            at_frontier && frontier == traversal_path[height - 1],
-                            data,
-                        )?;
+                        let (new_child, increment) = children[frontier]
+                            .extend_and_forget_internal::<H, Arity>(
+                                height - 1,
+                                &cur_pos,
+                                traversal_path,
+                                at_frontier && frontier == traversal_path[height - 1],
+                                data,
+                            )?;
+                        children[frontier] = new_child;
                         cnt += increment;
                         cur_pos += increment;
                         frontier += 1;
                     }
-                    *self = MerkleNode::Branch {
-                        value: digest_branch::<E, H, u64, T>(&children)?,
-                        children,
-                    };
-                    Ok(cnt)
+                    Ok((
+                        Arc::new(MerkleNode::Branch {
+                            value: digest_branch::<E, H, u64, T>(&children)?,
+                            children,
+                        }),
+                        cnt,
+                    ))
                 }
             },
             MerkleNode::Leaf { .. } => Err(PrimitivesError::ParameterError(
@@ -996,12 +1085,12 @@ where
 
 /// An owned iterator type for a merkle tree
 pub struct MerkleTreeIntoIter<E: Element, I: Index, T: NodeValue> {
-    stack: Vec<Box<MerkleNode<E, I, T>>>,
+    stack: Vec<Arc<MerkleNode<E, I, T>>>,
 }
 
 impl<E: Element, I: Index, T: NodeValue> MerkleTreeIntoIter<E, I, T> {
     /// Initialize an iterator
-    pub fn new(root: Box<MerkleNode<E, I, T>>) -> Self {
+    pub fn new(root: Arc<MerkleNode<E, I, T>>) -> Self {
         Self { stack: vec![root] }
     }
 }
@@ -1016,22 +1105,25 @@ where
 
     fn next(&mut self) -> Option<Self::Item> {
         while let Some(node) = self.stack.pop() {
-            match *node {
+            match node.as_ref() {
                 MerkleNode::Branch { value: _, children } => {
                     children
-                        .into_iter()
+                        .iter()
                         .rev()
                         .filter(|child| {
-                            matches!(**child, MerkleNode::Branch { .. } | MerkleNode::Leaf { .. })
+                            matches!(
+                                (**child).as_ref(),
+                                MerkleNode::Branch { .. } | MerkleNode::Leaf { .. }
+                            )
                         })
-                        .for_each(|child| self.stack.push(child));
+                        .for_each(|child| self.stack.push(child.clone()));
                 },
                 MerkleNode::Leaf {
                     value: _,
                     pos,
                     elem,
                 } => {
-                    return Some((pos, elem));
+                    return Some((pos.clone(), elem.clone()));
                 },
                 _ => {},
             }