+++ CHANGELOG.md
@@ -29,12 +29,13 @@ We take inspiration from [keep changelog](https://keepachangelog.com/en/1.0.0/) 
   - Switch from `crypto_box` to `chacha20poly1305` (with `crypto_kx` to establish shared secret) for AEAD.
   - Supports `--cfg curve25519_dalek_backend="u32_backend"` RUSTFLAGS to select Curve25519 backend.
   - Remove `Canonical(De)Serialize` on AEAD-related structs, and directly expose `serde::(De)Serialize` instead.
-- [#475](https://github.com/EspressoSystems/jellyfish/pull/475) (`jf-primitives`) Ergonomics changes for Merkle tree
+- [#475](https://github.com/EspressoSystems/jellyfish/pull/475) (`jf-primitives`) Ergonomics changes for Merkle tree.
   - Constructors are removed from trait definitions.
   - Remove the unnecessary trait bounds, for example `I: From<u64>`.
   - Restricting the index type for `AppendableMerkleTreeScheme` to be `u64`.
   - Add new `update_with()` and `remove()` interface for the universal Merkle tree.
   - Add new `iter()` interface for Merkle tree scheme, allows user to iterate through all elements that are in memory.
+- [#489](https://github.com/EspressoSystems/jellyfish/pull/489) (`jf-primitives`) Add non destructive universal Merkle tree.
 
 ### Fixed